requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiosignal==1.3.1
annotated-types==0.7.0
aspose-pdf==24.9.0
attrs==24.2.0
certifi==2024.8.30
coloredlogs==15.0.1
Deprecated==1.2.14
executor==23.2
fasteners==0.19
frozenlist==1.4.1
humanfriendly==10.0
idna==3.10
lxml==5.3.0
magic-filter==1.0.12
multidict==6.1.0
packaging==24.1
pikepdf==9.2.1
pillow==10.4.0
property-manager==3.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyPDF2==3.0.1
pyreadline3==3.5.4
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.12.2
verboselogs==1.7
wrapt==1.16.0
yarl==1.13.1
